Conflicts:
	CMakeLists.txt
	Model/essBuild/BilbaoWheel.cxx
	Model/essBuild/BilbaoWheelInnerStructure.cxx
	Model/essBuild/ButterflyModerator.cxx
	Model/essBuild/EdgeWater.cxx
	Model/essBuild/H2Wing.cxx
	Model/essBuild/MidWaterDivider.cxx
	Model/essBuild/WheelBase.cxx
	Model/essBuild/essVariables.cxx
	Model/essBuild/makeESS.cxx
	Model/essBuild/moderatorVariables.cxx
	Model/essBuildInc/BilbaoWheelInnerStructure.h
	Model/essBuildInc/EdgeWater.h
	Model/essBuildInc/F5Collimator.h
	Model/essBuildInc/H2Wing.h
	Model/essBuildInc/MidWaterDivider.h
	Model/essBuildInc/essVariables.h
	Model/essBuildInc/makeESS.h
	System/monte/DBModify.cxx

+++ Model/essBuildInc/ConicModerator.h
@@ -82,9 +82,9 @@ class ConicModerator : public constructSystem::ModBase
   /// Access to hydrogen region
   virtual int getMainBody() const { return HCell; }
 
-  virtual Geometry::Vec3D getSurfacePoint(const size_t,const size_t) const;
-  virtual int getLayerSurf(const size_t,const size_t) const;
-  virtual std::string getLayerString(const size_t,const size_t) const;
+  virtual Geometry::Vec3D getSurfacePoint(const size_t,const long int) const;
+  virtual int getLayerSurf(const size_t,const long int) const;
+  virtual std::string getLayerString(const size_t,const long int) const;
 
   virtual void createAll(Simulation&,const attachSystem::FixedComp&,
 			 const attachSystem::FixedComp*,const long int);
