+++ Model/essBuildInc/TaperedDiskPreMod.h
@@ -78,9 +78,9 @@ class TaperedDiskPreMod : public attachSystem::ContainedComp,
   virtual TaperedDiskPreMod* clone() const;
   virtual ~TaperedDiskPreMod();
 
-  virtual Geometry::Vec3D getSurfacePoint(const size_t,const size_t) const;
-  virtual int getLayerSurf(const size_t,const size_t) const;
-  virtual std::string getLayerString(const size_t,const size_t) const;
+  virtual Geometry::Vec3D getSurfacePoint(const size_t,const long int) const;
+  virtual int getLayerSurf(const size_t,const long int) const;
+  virtual std::string getLayerString(const size_t,const long int) const;
 
   /// total height of object
   double getZOffset() const { return zStep; }
